*: rebase and refactor out data structures in a separate package
This help in using the same packages across different projects,
so whenever we make a change it does not need to be copy pasted across
